backend/requirements.txt
fastapi==0.118.2
uvicorn==0.37.0
python-multipart==0.0.20
pillow==11.3.0
transformers==4.46.3
torch>=2.2.0
numpy<2
gpt4all==2.8.2